PscanR: An R Implementation for the `Pscan` Algorithm

`PscanR` provides an R implementation of the `Pscan` algorithm for transcription factor binding site motif analysis. The package is designed to work with Bioconductor data objects: it extends PFMatrix and PFMatrixList from the TFBSTools package for transcription factors binding profiles and employs Biostring DNAStringSet for promoter sequences. PscanR supports multithreading through the BiocParallel package. PscanR includes functions to build the background for a set of promoter sequences, to scan a foreground set of promoter sequences, and to visualize and plot the results.

Main functions

generate_psmatrixlist_from_background: Build background matrices from pre-computed files.

pscan: Pscan algorithm utilization on a set of gene promoters.

ps_results_table: Data visualization.

References

`Pscan` Web: http://159.149.160.88/pscan/

Zambelli F, Pesole G, Pavesi G. Pscan: finding over-represented transcription factor binding site motifs in sequences from co-regulated or co-expressed genes. Nucleic Acids Res. 2009 Jul;37(Web Server issue):W247-52. doi: 10.1093/nar/gkp464. Epub 2009 May 31. PMID: 19487240; PMCID: PMC2703934.

See also

Author

Federico Zambelli [cre], Diana Betelli [aut], Giulio Pavesi [cre]

Maintainer: Federico Zambelli <federico.zambelli@unimi.it>