Generate PSMatrixList from Background data and JASPAR Matrix

This function retrieves background data for a specified organism and promoter region, and then fetches a corresponding matrix set from a specified JASPAR version. It returns a `PSMatrixList` object with background statistics.

generate_psmatrixlist_from_background(
  JASPAR_matrix,
  org,
  prom_reg,
  assembly = character()
)

Arguments

JASPAR_matrix

A character string specifying the JASPAR database version. You can choose between 'JASPAR2020', 'JASPAR2022', or 'JASPAR2024' (non case sensitive).

org

A string representing the organism acronym. Accepted values are: `hs` (Homo sapiens), `mm` (Mus musculus), `at` (Arabidopsis thaliana), `sc` (Saccharomyces cerevisiae), `dm` (Drosophila melanogaster).

prom_reg

A numeric vector of two integers as `[upstream, downstream]` representing the promoter region relative to the TSS. Allowed combinations:

200 base pairs upstream and 50 downstream base pair from the TSS (c(-200, 50))

450 bp upstream and 50 downstream (c(-450, 50))

500 bp upstream and 0 downstream (c(-500, 0))

950 bp upstream and 50 downstream (c(-950, 50))

1000 bp upstream and 0 downstream (c(-1000, 0))

assembly

A string representing the assembly version for Human or mouse. For `"hs"` you can choose between `"hg38"` or the latest `"hs1"`. For `"mm"` you can specify `"mm10"` or `"mm39"`. Default is character().

Value

A `PSMatrixList` object with background-scored motif matrices.

Details

The background files are downloaded from the GitHub repository: https://github.com/dianabetelli/PscanR_backgrounds This function automatically fetches the appropriate background file and combines it with the specified JASPAR matrix collection to create a `PSMatrixList` with background statistics.

Examples

bg_matrices <- generate_psmatrixlist_from_background('Jaspar2020', 'hs', 
                                                     c(-200,50), 'hg38')
bg_matrices
#> PSMatrixList of length 746
#> names(746): MA0004.1 MA0006.1 MA0019.1 MA0029.1 ... MA0748.2 MA0528.2 MA0609.2
bg_matrices[[4]]
#> An object of class PSMatrix
#> ID: MA0029.1
#> Name: Mecom
#> Matrix Class: C2H2 zinc finger factors
#> strand: +
#> Tags: 
#> $alias
#> [1] "MDS1-EVI1,PRDM3"
#> 
#> $description
#> [1] "MDS1 and EVI1 complex locus MDS1,EVI1"
#> 
#> $family
#> [1] "Factors with multiple dispersed zinc fingers"
#> 
#> $medline
#> [1] "8321231"
#> 
#> $symbol
#> [1] "MECOM"
#> 
#> $tax_group
#> [1] "vertebrates"
#> 
#> $tfbs_shape_id
#> [1] "36"
#> 
#> $type
#> [1] "SELEX"
#> 
#> $collection
#> [1] "CORE"
#> 
#> $species
#>          10090 
#> "Mus musculus" 
#> 
#> $acc
#> [1] "A4QPC8"
#> 
#> Background: 
#>    A    C    G    T 
#> 0.25 0.25 0.25 0.25 
#> Matrix: 
#>         [,1]       [,2]       [,3]        [,4]       [,5]        [,6]
#> A -0.6768981 -0.3259156 -4.6443909 -0.02927038 -3.0960407 -0.02927038
#> C -2.5152777 -3.0960407 -3.0960407 -4.64439090 -1.0058306 -4.64439090
#> G -1.4992812 -2.5152777 -0.1054397 -4.64439090 -4.6443909 -4.64439090
#> T -1.6730272 -1.8834438 -3.0960407 -4.64439090 -0.5457347 -4.64439090
#>          [,7]        [,8]        [,9]      [,10]      [,11]      [,12]
#> A -0.06662998 -4.64439090 -0.02927038 -4.6443909 -0.1458166 -0.1878926
#> C -4.64439090 -4.64439090 -4.64439090 -2.1502676 -3.0960407 -4.6443909
#> G -3.09604068 -0.02927038 -4.64439090 -4.6443909 -4.6443909 -1.8834438
#> T -4.64439090 -4.64439090 -4.64439090 -0.1458166 -2.5152777 -4.6443909
#>       [,13]      [,14]
#> A -1.499281 -0.6091675
#> C -1.351301 -1.4992812
#> G -1.351301 -2.1502676
#> T -1.351301 -2.1502676
#> 
#> Pscan Background Average: 0.7090597
#> Pscan Foreground (your sample) Average: NA
#> Pscan Backgroun Stdev: 0.06468208
#> Pscan Background Size: 39438
#> Pscan Foreground (your sample) Size: NA
#> Pscan Zscore: NA
#> Pscan pvalue (z-test): NA