Build background matrices in parallel

This function generates a background probability profile matrix list, used to assess if a transcription factor binding motif is appearing more often than by chance in the pscan() input sequences, using the BiocParallel framework for parallel execution. It computes background scores for each motif matrix, collected, for example, from the JASPAR database, by scanning them against a set of regulatory sequences (e.g., gene promoters) using the `ps_scan` function. Note that the promoter region analysed may vary.

ps_build_bg(
  x,
  pfms,
  BPPARAM = bpparam(),
  BPOPTIONS = bpoptions(),
  fullBG = FALSE
)

Arguments

x

A `DNAStringSet` object (see Biostrings package) containing the set of all the regulatory sequences of the organism of study (e.g., gene promoters retrieved from a `TxDb` object). These sequences are the target for background scanning.

pfms

A `PFMatrixlist` object containing position frequency matrices (PFM) representing transcription factor binding preferences. Those are sourced from databases such as JASPAR.

BPPARAM

Parallelization parameter passed to `bplapply` function from the `BiocParallel` package. This parameter defines the parallel processing settings, including the number of cores or workers that the function uses for computation. See `BiocParallel` package for more details.

BPOPTIONS

Optional configuration settings passed to `bplapply` function from the `BiocParallel` package. This can include additional options to control the behavior of parallel execution. See `BiocParallel` documentation for more details.

fullBG

Logical. Default is FALSE. When set to TRUE, it creates a mapping between all sequence names in the organism of study and the corresponding names retained after applying the unique() function. For example, if multiple identical sequences exist (e.g., ID1, ID2, ID3, and ID4), and unique() retains only ID2, the mapping will associate each original name with its unique counterpart (ID1 → ID2, ID2 → ID2, ID3 → ID2, ID4 → ID2). This vector is stored in the transcriptIDLegend slot of the PSMatrixList and helps generate a complete background PSMatrixList, improving computational efficiency for the pscan_fullBG() function. In addition, it retrieves for each PSM in the PSMatrixList output all the background metrics relative to hits score, position, strand, and oligonucleotide sequence of the regulatory sequences scanned with the PWM.

Value

A `PSMatrixList` object, containing each motif matrix from `pfms`, background-scored against the sequences in `x`.

Details

This function validates input types and removes duplicated sequences from `x` to avoid redundant computations. It also remove all the sequences with an N content above 50 The motif matrices are background scored by the `ps_scan` function in parallel. The output of this function can be stored in a .txt file by the `ps_write_bg_to_file()` function. In this case, only information about the background mean and standard deviation are stored for each matrix. These metrics will be later used for the computation of z-score by the pscan function.

If a full background PSMatrixList is required, that includes all the background scores for each oligonucleotide hits, their position, strand and names, the user should store the output by the save() function and set the fullBG flag as TRUE. Note that this process will generate a very large file, that can reach several gigabytes in size.

Note: Currently, it is not possible to use this function with the Arabidopsis thaliana organism due to the transcript nomenclature used. The function is designed to handle transcript names without version extensions (e.g., NM_30287 instead of NM_30287.1). However, in Arabidopsis thaliana, the format typically follows a gene.transcript pattern (e.g., AT1G01010.1). This causes the function to truncate the transcript identifier, effectively removing essential transcript-level information. As a result, the function may return incorrect results for this organism. This is valid for any organism for which this type of format is used.

This function uses example datasets located in the `extdata/` directory for demonstration purposes only. These files are not part of the core data used by the function. They can be accessed using `system.file()` as shown in the examples.

See also

Examples

# Load the example dataset for promoter sequences (hg38 assembly, 
# -200 +50 bp in respect to the TSS).
file_path <- system.file("extdata", "prom_seq.rds", package = "PscanR")
prom_seq <- readRDS(file_path)
prom_seq <- prom_seq[1:10]

# Load the example dataset for JASPAR2020 matrices collection for 
# vertebrates.
J2020_path <- system.file("extdata", "J2020.rds", package = "PscanR")
J2020 <- readRDS(J2020_path)

# Generate the background-scored motif matrices
bg_matrices <- ps_build_bg(prom_seq, J2020, 
                           BPPARAM = BiocParallel::SnowParam(1))
# Use BiocParallel::MulticoreParam() for Unix like systems. 
bg_matrices
#> PSMatrixList of length 746
#> names(746): MA0004.1 MA0006.1 MA0019.1 MA0029.1 ... MA0748.2 MA0528.2 MA0609.2
bg_matrices[[1]]
#> An object of class PSMatrix
#> ID: MA0004.1
#> Name: Arnt
#> Matrix Class: Basic helix-loop-helix factors (bHLH)
#> strand: +
#> Tags: 
#> $alias
#> [1] "HIF-1beta,bHLHe2"
#> 
#> $description
#> [1] "aryl hydrocarbon receptor nuclear translocator"
#> 
#> $family
#> [1] "PAS domain factors"
#> 
#> $medline
#> [1] "7592839"
#> 
#> $remap_tf_name
#> [1] "ARNT"
#> 
#> $symbol
#> [1] "ARNT"
#> 
#> $tax_group
#> [1] "vertebrates"
#> 
#> $tfbs_shape_id
#> [1] "11"
#> 
#> $type
#> [1] "SELEX"
#> 
#> $unibind
#> [1] "1"
#> 
#> $collection
#> [1] "CORE"
#> 
#> $species
#>          10090 
#> "Mus musculus" 
#> 
#> $acc
#> [1] "P53762"
#> 
#> Background: 
#>    A    C    G    T 
#> 0.25 0.25 0.25 0.25 
#> Matrix: 
#>         [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> A -1.5998685 -0.08004271 -4.64439090 -4.64439090 -4.64439090 -4.64439090
#> C -0.2499417 -4.64439090 -0.02927038 -4.64439090 -4.64439090 -4.64439090
#> G -4.6443909 -2.85263143 -4.64439090 -0.02927038 -4.64439090 -0.02927038
#> T -4.6443909 -4.64439090 -4.64439090 -4.64439090 -0.02927038 -4.64439090
#> 
#> Pscan Background Average: 0.8205067
#> Pscan Foreground (your sample) Average: NA
#> Pscan Backgroun Stdev: 0.08426005
#> Pscan Background Size: 10
#> Pscan Foreground (your sample) Size: NA
#> Pscan Zscore: NA
#> Pscan pvalue (z-test): NA

# Example for full-background generation
full_bg_matrices <- ps_build_bg(prom_seq, J2020, 
                           BPPARAM = BiocParallel::SnowParam(1), 
                           fullBG = TRUE)
                           
full_bg_matrices[[1]]
#> An object of class PSMatrix
#> ID: MA0004.1
#> Name: Arnt
#> Matrix Class: Basic helix-loop-helix factors (bHLH)
#> strand: +
#> Tags: 
#> $alias
#> [1] "HIF-1beta,bHLHe2"
#> 
#> $description
#> [1] "aryl hydrocarbon receptor nuclear translocator"
#> 
#> $family
#> [1] "PAS domain factors"
#> 
#> $medline
#> [1] "7592839"
#> 
#> $remap_tf_name
#> [1] "ARNT"
#> 
#> $symbol
#> [1] "ARNT"
#> 
#> $tax_group
#> [1] "vertebrates"
#> 
#> $tfbs_shape_id
#> [1] "11"
#> 
#> $type
#> [1] "SELEX"
#> 
#> $unibind
#> [1] "1"
#> 
#> $collection
#> [1] "CORE"
#> 
#> $species
#>          10090 
#> "Mus musculus" 
#> 
#> $acc
#> [1] "P53762"
#> 
#> Background: 
#>    A    C    G    T 
#> 0.25 0.25 0.25 0.25 
#> Matrix: 
#>         [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> A -1.5998685 -0.08004271 -4.64439090 -4.64439090 -4.64439090 -4.64439090
#> C -0.2499417 -4.64439090 -0.02927038 -4.64439090 -4.64439090 -4.64439090
#> G -4.6443909 -2.85263143 -4.64439090 -0.02927038 -4.64439090 -0.02927038
#> T -4.6443909 -4.64439090 -4.64439090 -4.64439090 -0.02927038 -4.64439090
#> 
#> Pscan Background Average: 0.8205067
#> Pscan Foreground (your sample) Average: NA
#> Pscan Backgroun Stdev: 0.08426005
#> Pscan Background Size: 10
#> Pscan Foreground (your sample) Size: NA
#> Pscan Zscore: NA
#> Pscan pvalue (z-test): NA