Apply background parameters to position frequency matrices

Associates background parameters from a `data.frame` with a list of position frequency matrices (PFMs) and returns the updated matrices as a `PSMatrixList` object.

ps_build_bg_from_table(x, pfms)

Arguments

x

A `data.frame` containing background parameters for each motif. The data frame should include the following columns:

`BG_SIZE`: Background set size.

`BG_MEAN`: Mean score of background hits.

`BG_STDEV`: Standard deviation of background hits score.

The row names of `x` should correspond to the identifiers of the motifs in `pfms`.

pfms

A `PFMatrixList` object containing position frequency matrices representing transcription factor binding preferences, obtained, for example, from the JASPAR database. The number of elements in `pfms` should match the number of row of `x`, and their identifiers should correspond.

Value

A `PSMatrixList` object containing each motif from `pfms`, enriched with background scoring parameters from `x`.

Details

This function:

Validates the input type. `x` must be a `data.frame`.

Converts each elements of pfms into `PSMatrix` objects.

A warning is issued if the number of elements in `pfms` doesn't match the number of row of `x`.

Note: This function does not compute new background scores — it assigns existing background parameters to each motif, assuming they were previously computed.

This function uses example datasets located in the `extdata/` directory for demonstration purposes only. These files are not part of the core data used by the function. They can be accessed using `system.file()` as shown in the examples.

Examples

# create the `data.frame`
background_data <- data.frame(
BG_SIZE = c(500, 450, 480),
BG_MEAN = c(0.3, 0.25, 0.35),
BG_STDEV = c(0.05, 0.07, 0.06)
)

# Retrieve motif matrices for vertebrates from JASPAR2020
J2020_path <- system.file("extdata", "J2020.rds", package = "PscanR")
J2020 <- readRDS(J2020_path)
J2020_subset <- J2020[1:3] # match the number of rows in `backgound_data`

rownames(background_data) <- c("MA0004.1", "MA0006.1", "MA0019.1")

# Generate background-scored motif matrices
bg_matrices <- ps_build_bg_from_table(background_data, J2020_subset)
bg_matrices
#> PSMatrixList of length 3
#> names(3): MA0004.1 MA0006.1 MA0019.1
bg_matrices[[1]]
#> An object of class PSMatrix
#> ID: MA0004.1
#> Name: Arnt
#> Matrix Class: Basic helix-loop-helix factors (bHLH)
#> strand: +
#> Tags: 
#> $alias
#> [1] "HIF-1beta,bHLHe2"
#> 
#> $description
#> [1] "aryl hydrocarbon receptor nuclear translocator"
#> 
#> $family
#> [1] "PAS domain factors"
#> 
#> $medline
#> [1] "7592839"
#> 
#> $remap_tf_name
#> [1] "ARNT"
#> 
#> $symbol
#> [1] "ARNT"
#> 
#> $tax_group
#> [1] "vertebrates"
#> 
#> $tfbs_shape_id
#> [1] "11"
#> 
#> $type
#> [1] "SELEX"
#> 
#> $unibind
#> [1] "1"
#> 
#> $collection
#> [1] "CORE"
#> 
#> $species
#>          10090 
#> "Mus musculus" 
#> 
#> $acc
#> [1] "P53762"
#> 
#> Background: 
#>    A    C    G    T 
#> 0.25 0.25 0.25 0.25 
#> Matrix: 
#>         [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> A -1.5998685 -0.08004271 -4.64439090 -4.64439090 -4.64439090 -4.64439090
#> C -0.2499417 -4.64439090 -0.02927038 -4.64439090 -4.64439090 -4.64439090
#> G -4.6443909 -2.85263143 -4.64439090 -0.02927038 -4.64439090 -0.02927038
#> T -4.6443909 -4.64439090 -4.64439090 -4.64439090 -0.02927038 -4.64439090
#> 
#> Pscan Background Average: 0.3
#> Pscan Foreground (your sample) Average: NA
#> Pscan Backgroun Stdev: 0.05
#> Pscan Background Size: 500
#> Pscan Foreground (your sample) Size: NA
#> Pscan Zscore: NA
#> Pscan pvalue (z-test): NA