Extract background statistics from a `PSMatrixList` object

Extracts background statistics (size, mean, and standard deviation) from a `PSMatrixList` of Position Weight Matrices representing transcription factor binding preferences, and returns a data frame containing these metrics.

ps_get_bg_table(pfms)

Arguments

pfms

A `PSMatrixList` of Position Weight Matrices representing transcription factor binding preferences . Each element must be a `PSMatrix` object or coercible to one.

Value

A `data.frame` with one row per matrix (corresponding to a transcription factor) in `pfms`. The row names will match the matrix identifiers.

Columns:

`BG_SIZE`: An integer vector containing the background set size for each PSM.

`BG_MEAN`: A numeric vector containing the mean of the background hits score for each PSM.

`BG_STDEV`: A numeric vector containing the standard deviation of the background hits score for each PSM.

Details

This function is useful for exporting background-scored matrices to a tabular format. The resulting `data.frame` can be saved and later reapplied to PFMs using `ps_build_bg_from_table()`.

This function uses example datasets located in the `extdata/` directory for demonstration purposes only. These files are not part of the core data used by the function. They can be accessed using `system.file()` as shown in the examples.

See also

Examples

# Retrieve motif matrices for vertebrates from JASPAR2020
J2020_path <- system.file("extdata", "J2020.rds", package = "PscanR")
J2020 <- readRDS(J2020_path)
J2020_subset <- J2020[1:3] # match the number of rows in `backgound_data`

# create the `data.frame`
background_data <- data.frame(
BG_SIZE = c(500, 450, 480),
BG_MEAN = c(0.3, 0.25, 0.35),
BG_STDEV = c(0.05, 0.07, 0.06)
)

rownames(background_data) <- c("MA0004.1", "MA0006.1", "MA0019.1")

# Generate background-scored motif matrices
bg_matrices <- ps_build_bg_from_table(background_data, J2020_subset)
bg_table <- ps_get_bg_table(bg_matrices)
bg_table 
#>          BG_SIZE BG_MEAN BG_STDEV
#> MA0004.1     500    0.30     0.05
#> MA0006.1     450    0.25     0.07
#> MA0019.1     480    0.35     0.06