Get Matched Oligonucleotide Sequences

Retrieves the oligonucleotide sequences that match the motif in the foreground sequences, returning a character vector.

# S4 method for class 'PSMatrix'
ps_hits_oligo(x, withDimnames = TRUE)

Arguments

x

A `PSMatrix` object, typically the result of the `Pscan` algorithm.

withDimnames

Logical, whether to include dimension names in the output, if they exist in the object. Default set to `TRUE`.

Value

A character vector containing the sequences of motif matches (oligonucleotides) in the foreground set.

Examples

pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
ps_hits_oligo(pfm1)
#> NM_001078.4 NM_000639.3 NM_004433.5 NM_001877.5 NM_001561.6 NM_003039.3 
#>    "AACTTG"    "AAAGTG"    "CAAGTG"    "CACGTG"    "AGCCTG"    "CACGCG" 
#> NM_000952.5 NM_001993.5 NM_014624.4 NM_000567.3 NM_001778.4 NM_003005.4 
#>    "CACCTT"    "CAAGTG"    "AATGTG"    "CACCTG"    "CAAGTG"    "AATGTG" 
#> NM_000450.2 NM_000963.4 NM_000577.5 NM_004591.3 NM_015727.3 NM_000575.5 
#>    "ATCGTG"    "CACGTC"    "CGGGTG"    "CATGTG"    "CTCGTG"    "CACGTA" 
#> NM_000576.3 NM_000579.4 NM_002852.4 NM_006225.4 NM_005191.4 NM_000584.4 
#>    "AACTTG"    "CATGTG"    "CACTTG"    "CGCGTC"    "CTCGTG"    "ATCGTG" 
#> NM_003998.4 NM_002994.5 NM_000586.4 NM_005429.5 NM_002199.4 NM_000758.4 
#>    "CGCGTG"    "CACGAG"    "CACATG"    "CACGCT"    "CTCGCG"    "CATGTG" 
#> NM_002188.3 NM_002198.3 NM_000590.2 NM_004355.4 NM_002460.4 NM_004233.4 
#>    "CACTTG"    "GACGTG"    "CAAGTG"    "CAAGTG"    "CACGCC"    "CTCGTG" 
#> NM_000595.4 NM_000594.4 NM_002800.5 NM_006290.4 NM_000914.5 NM_000367.5 
#>    "CACGCT"    "AAGGTG"    "CACGTT"    "CACGTG"    "CACGCT"    "AACGAG" 
#> NM_003897.4 NM_009588.1 NM_172197.3 NM_000593.6 NM_000636.4 NM_000600.5 
#>    "AACGTG"    "CAGGTG"    "AGGGTG"    "CACTTT"    "CCCGTG"    "AATGTG" 
#> NM_002787.5 NM_006528.4 
#>    "CATGTG"    "CCCGTG"