Get Motif Hit Positions in a Backrgound Dataset

Retrieves the positions of hits stored in a `PSMatrix` object computed on the background dataset (all the promoter in a specific organism). These positions indicate where the motifs are located in each promoter sequence.

# S4 method for class 'PSMatrix'
ps_hits_pos_bg(x, withDimnames = TRUE)

Arguments

x

A `PSMatrix` object.

withDimnames

Logical, whether to include dimension names in the output, if they exist in the object. Default set to `TRUE`.

Value

A named integer vector where names correspond to promoter sequence identifiers, and values represent the motif hit positions.

Details

This method is specifically designed for background datasets, where motif hit positions are precomputed for all promoter sequences. The function extracts the position values from the ps_hits_pos_bg slot and, if applicable, assigns sequence names using .ps_bg_seq_names().

These background metrics are particularly useful in motif enrichment analyses, as they allow pscan() to compare foreground promoter sequences against a reference distribution without the need for recomputation.

Examples

full_pfm1_path <- system.file("extdata", 
                              "full_pfm1.rds", 
                              package = "PscanR")
full_pfm1 <- readRDS(full_pfm1_path)
ps_hits_pos_bg(full_pfm1)
#>    NR_046018.2    NR_036051.1 NM_001005484.2    NR_148357.1    NR_187359.1 
#>             45             35            147              1             42 
#>    NR_103536.1    NR_047526.1    NR_047525.1 NM_001385640.1    NM_152486.4 
#>            144            165            100            181            127 
#>    NM_198317.3 NM_001160184.2    NM_005101.4 NM_001305275.2 NM_001364727.2 
#>            126             64            151             82             57 
#>    NR_148960.1    NR_038869.1    NR_029639.1    NR_029834.1    NR_029957.1 
#>            241             61             19            191             38 
#> NM_001130045.2    NM_153254.3    NM_080605.4    NR_185846.1 NM_001130413.4 
#>              5            192             74             40             87 
#> NM_001346116.2 NM_001029885.2    NM_152228.3    NR_015434.1    NR_170330.2 
#>             52            104            144             60              1 
#>    NR_187360.1 NM_001146685.2    NM_022834.5 NM_001039211.3    NM_031921.6 
#>             22             66             83             65            156 
#> NM_001317238.2 
#>            202