Get Table of Motif Hits

Creates a data frame summarizing the motif hits from a `PSMatrix` object, including the motif hit score, position, strand, and corresponding oligo sequence. The resulting table is sorted by motif score (descending). When the score is the same between two sequences, they are sorted by ascending position.

# S4 method for class 'PSMatrix'
ps_hits_table(x, pos_shift = 0L, withDimnames = TRUE)

Arguments

x

An object of class `PSMatrix`. Should contain the following slots:

`ps_hits_score`: a numeric vector of hit scores.

`ps_hits_strand`: a character vector of strand information (`-` and `+`).

`ps_hits_oligo`: a character vector of oligo sequences.

pos_shift

Integer. Value for which the position gets shifted. Default is `0`, meaning no shift.

withDimnames

Logical, whether to include dimension names in the output, if they exist in the object. Default set to `TRUE`.

Value

A `data.frame` ordered by decreasing score value, with the following columns:

`SCORE`: the motif hit score

`POS`: the position of the motif hit

`STRAND`: the strand orientation for each hit

`OLIGO`: the oligo sequence corresponding to each hit

Row names correspond to the sequences name.

Examples

pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
ps_hits_table(pfm1)
#>                 SCORE POS STRAND  OLIGO
#> NM_006290.4 1.0000000 123      + CACGTG
#> NM_001877.5 1.0000000 156      + CACGTG
#> NM_003897.4 0.9507672  23      + AACGTG
#> NM_002800.5 0.9507672  89      - CACGTT
#> NM_003998.4 0.8988818 112      + CGCGTG
#> NM_003039.3 0.8988818 244      - CACGCG
#> NM_005429.5 0.8496490 142      - CACGCT
#> NM_000595.4 0.8496490 147      - CACGCT
#> NM_000914.5 0.8496490 236      - CACGCT
#> NM_000575.5 0.8397314 143      - CACGTA
#> NM_002198.3 0.8397314 229      + GACGTG
#> NM_000963.4 0.8397314 234      - CACGTC
#> NM_000636.4 0.8335351   1      + CCCGTG
#> NM_005191.4 0.8335351  33      + CTCGTG
#> NM_006528.4 0.8335351 106      + CCCGTG
#> NM_015727.3 0.8335351 181      + CTCGTG
#> NM_002994.5 0.8335351 189      - CACGAG
#> NM_004233.4 0.8335351 213      + CTCGTG
#> NM_000590.2 0.8316834  53      + CAAGTG
#> NM_004433.5 0.8316834  58      + CAAGTG
#> NM_000758.4 0.8316834  77      + CATGTG
#> NM_001993.5 0.8316834  80      + CAAGTG
#> NM_000586.4 0.8316834  80      + CACATG
#> NM_000567.3 0.8316834 104      + CACCTG
#> NM_001778.4 0.8316834 114      + CAAGTG
#> NM_004355.4 0.8316834 125      + CAAGTG
#> NM_009588.1 0.8316834 128      + CAGGTG
#> NM_004591.3 0.8316834 131      + CATGTG
#> NM_002188.3 0.8316834 160      + CACTTG
#> NM_000579.4 0.8316834 180      + CATGTG
#> NM_002787.5 0.8316834 181      + CATGTG
#> NM_002852.4 0.8316834 231      + CACTTG
#> NM_000450.2 0.7843023  70      + ATCGTG
#> NM_000584.4 0.7843023 105      + ATCGTG
#> NM_000367.5 0.7824506   1      + AACGAG
#> NM_000952.5 0.7824506  15      - CACCTT
#> NM_000593.6 0.7824506  26      - CACTTT
#> NM_000594.4 0.7824506  59      + AAGGTG
#> NM_014624.4 0.7824506  73      + AATGTG
#> NM_000639.3 0.7824506  78      + AAAGTG
#> NM_001078.4 0.7824506  79      + AACTTG
#> NM_000576.3 0.7824506 103      + AACTTG
#> NM_000600.5 0.7824506 123      + AATGTG
#> NM_003005.4 0.7824506 139      + AATGTG
#> NM_002460.4 0.7386132  36      - CACGCC
#> NM_002199.4 0.7305652  37      - CTCGCG
#> NM_006225.4 0.7305652  81      + CGCGTC
#> NM_000577.5 0.7305652 165      + CGGGTG
#> NM_172197.3 0.6813324 154      + AGGGTG
#> NM_001561.6 0.6813324 206      + AGCCTG