Import background-scored matrices from a file

Reconstructs a `PSMatrixList` object containing background-scored motif matrices, using a previously computed background dataset stored in a file. The input file must have been generated by the `ps_write_bg_to_file()` function from a background `PSMatrixList` object.

ps_retrieve_bg_from_file(file, pfms)

Arguments

file

A character string with the path to the input file. This file contains the score distribution statistics for the frequency matrices computed on a background set of regulatory sequences. The file should be in tabular format, where the first column contains the frequency matrix identifiers (row names) and subsequent columns contain the background statistics (e.g., BG_SIZE, BG_MEAN, BG_STDEV). Background statistics are generated with the `ps_build_bg` function and can be written to a file with `ps_write_bg_to_file`.

pfms

A `PFMatrixList` object containing position frequency matrices representing transcription factor binding preferences, obtained, for example, from the JASPAR database. The provided list must correspond to the same frequency matrix list used to build the background stored in `file`.

Value

A `PSMatrixList` object, containing each motif matrix from `pfms`, background-scored using the values provided in `file`. See `ps_build_bg_from_table` for more details.

Details

This function:

Validates the input type: must be a character string.

Reads the input file into a `data.frame`, with row names taken from the first column.

Calls `ps_build_bg_from_table` with the created `data.frame` and `pfms` as input.

This function uses example datasets located in the `extdata/` directory for demonstration purposes only. These file are not part of the core data used by the function. They can be accessed using `system.file()` as shown in the examples.

Other background datasets are aviable at the public repository PscanR_background on GitHub: https://github.com/dianabetelli/PscanR_backgrounds See vignettes for further details on the type of background available.

See also

Examples

# Load a background information file
file_path <- system.file("extdata", "J2020_hg38_200u_50d_UCSC.psbg.txt", 
                         package = "PscanR")

# Load the example dataset for JASPAR2020 matrices collection for 
# vertebrates.
J2020_path <- system.file("extdata", "J2020.rds", package = "PscanR")
J2020 <- readRDS(J2020_path)

# Generate the background-scored motif matrices from file
bg_matrices <- ps_retrieve_bg_from_file(file_path, J2020)
bg_matrices
#> PSMatrixList of length 746
#> names(746): MA0004.1 MA0006.1 MA0019.1 MA0029.1 ... MA0748.2 MA0528.2 MA0609.2
bg_matrices[[2]]
#> An object of class PSMatrix
#> ID: MA0006.1
#> Name: Ahr::Arnt
#> Matrix Class: Basic helix-loop-helix factors (bHLH)
#> strand: +
#> Tags: 
#> $alias
#> [1] "-"
#> 
#> $comment
#> [1] "dimer"
#> 
#> $description
#> [1] "-"
#> 
#> $family
#> [1] "PAS domain factors"
#> 
#> $medline
#> [1] "7592839"
#> 
#> $remap_tf_name
#> [1] "AHR"
#> 
#> $symbol
#> [1] "-"
#> 
#> $tax_group
#> [1] "vertebrates"
#> 
#> $tfbs_shape_id
#> [1] "13"
#> 
#> $type
#> [1] "SELEX"
#> 
#> $unibind
#> [1] "1"
#> 
#> $collection
#> [1] "CORE"
#> 
#> $species
#>          10090 
#> "Mus musculus" 
#> 
#> $acc
#> [1] "P30561" "P53762"
#> 
#> Background: 
#>    A    C    G    T 
#> 0.25 0.25 0.25 0.25 
#> Matrix: 
#>         [,1]        [,2]        [,3]        [,4]        [,5]        [,6]
#> A -2.0417012 -4.64439090 -4.64439090 -4.64439090 -4.64439090 -4.64439090
#> C -1.1082742 -4.64439090 -0.07139961 -4.64439090 -4.64439090 -4.64439090
#> G -2.4107987 -0.07139961 -4.64439090 -0.07139961 -4.64439090 -0.02927038
#> T -0.7977957 -3.00216316 -3.00216316 -3.00216316 -0.02927038 -4.64439090
#> 
#> Pscan Background Average: 0.8806266
#> Pscan Foreground (your sample) Average: NA
#> Pscan Backgroun Stdev: 0.07161552
#> Pscan Background Size: 39438
#> Pscan Foreground (your sample) Size: NA
#> Pscan Zscore: NA
#> Pscan pvalue (z-test): NA