Perform a Scan of DNA Sequences Using a `PSMatrix` Object

This method performs a scan of DNA sequences using a `PSMatrix` object to identify potential hits based on the matrix score. The scan is performed on both strands (the direct and reverse complement strands) of the sequences to ensure that no potential binding sites are missed.

# S4 method for class 'PSMatrix'
ps_scan(x, seqs, BG = FALSE, use_full_BG = FALSE, fullBG = FALSE)

Arguments

x

A `PSMatrix` object representing the motif and related statistics to be used for scanning DNA sequences.

seqs

`A DNAstringSet` object representing the sequences to be scanned for motif occurrences.

BG

A logical value indicating whether to calculate background statistics. Default is set to `FALSE`.

use_full_BG

A logical value (default is `FALSE`). If `TRUE`, the method assumes that the PSMatrix represent a special "full-background" set. In this case, the `seqs` parameter should be a vector of sequence names (instead of actual DNA sequences). The function handles this by matching sequence names against precomputed background data in the `PSMatrix`.

fullBG

Logical, default is `FALSE`. If `TRUE` it computes a series of computation for the generation of a "full-background", a special case of background that retains all the background hits score, position, strand, and oligonucleotide sequence for each regulatory sequence scanned with a PWM.

Value

A `PSMatrix` object, with updated information about the motif hits in the sequences. This includes the positions, strands, scores, and oligos (sequence motifs) where the hits occurred.

Details

This method is designed to handle different types of sequence scanning scenarios. It can handle both regular DNA sequences (using a `DNAStringSet`) and specialized background sequence sets (either using a background flag or a full-background flag).

If `use_full_BG` is set to `TRUE`, the function assumes that `seqs` contains sequence identifiers rather than the sequences themselves. In this case, the method matches the sequence names to those in the `PSMatrix`'s background hit data and retrieves the corresponding binding information (score, strand, position, oligo).

When `use_full_BG` is set to `FALSE`, the function scan both the forward and reverse complement strands of the sequences to ensure all potential binding sites are detected. Optionally, the background statistics (background average and standard deviation) can be computed and used during scanning when `BG` is set to `TRUE`.

Examples

BG_matrices <- generate_psmatrixlist_from_background('Jaspar2020', 'hs', 
                                                     c(-200,50),'hg38')
file_path <- system.file("extdata", "prom_seq.rds", package = "PscanR")
prom_seq <- readRDS(file_path)
prom_seq <- prom_seq[1:25]

scanned_result <- ps_scan(BG_matrices[[3]], prom_seq)
scanned_result
#> An object of class PSMatrix
#> ID: MA0019.1
#> Name: Ddit3::Cebpa
#> Matrix Class: Basic leucine zipper factors (bZIP)
#> strand: +
#> Tags: 
#> $alias
#> [1] "-"
#> 
#> $comment
#> [1] "dimer between Ddit3 and Cebpa"
#> 
#> $description
#> [1] "-"
#> 
#> $family
#> [1] "C/EBP-related factors"
#> 
#> $medline
#> [1] "8657121"
#> 
#> $symbol
#> [1] "-"
#> 
#> $tax_group
#> [1] "vertebrates"
#> 
#> $tfbs_shape_id
#> [1] "26"
#> 
#> $type
#> [1] "SELEX"
#> 
#> $collection
#> [1] "CORE"
#> 
#> $species
#>               10116 
#> "Rattus norvegicus" 
#> 
#> $acc
#> [1] "Q62857" "P05554"
#> 
#> Background: 
#>    A    C    G    T 
#> 0.25 0.25 0.25 0.25 
#> Matrix: 
#>        [,1]      [,2]       [,3]        [,4]        [,5]       [,6]        [,7]
#> A -1.036249 -1.270047 -0.7909753 -4.64439090 -4.64439090 -2.2234531 -0.05498535
#> C -1.702655 -1.702655 -2.4819524 -3.37347861 -4.64439090 -0.1945259 -3.37347861
#> G -1.185893 -1.036249 -0.9690644 -4.64439090 -0.05498535 -4.6443909 -4.64439090
#> T -1.848048 -1.702655 -2.4819524 -0.05498535 -3.37347861 -2.8314890 -4.64439090
#>         [,8]       [,9]     [,10]      [,11]      [,12]
#> A -0.1084883 -4.6443909 -1.036249 -2.2234531 -4.6443909
#> C -2.8314890 -1.8480481 -0.846887 -0.5504738 -0.4297972
#> G -3.3734786 -4.6443909 -2.018237 -1.4631371 -1.8480481
#> T -4.6443909 -0.1945259 -2.481952 -2.4819524 -1.7026546
#> 
#> Pscan Background Average: 0.7960527
#> Pscan Foreground (your sample) Average: 0.79835
#> Pscan Backgroun Stdev: 0.04700296
#> Pscan Background Size: 39438
#> Pscan Foreground (your sample) Size: 25
#> Pscan Zscore: 0.2443804
#> Pscan pvalue (z-test): 0.4034681