Retrieve Sequence Names (Generic Function)

`ps_seq_names` is a **generic function** that extracts the sequence names or identifiers from an object containing promoter sequence data.

ps_seq_names(x, ...)

Arguments

x

An object containing sequence information.

...

Additional arguments passed to specific methods.

Value

If `x` is a `PSMatrix` object: a **character vector** of sequence names corresponding to the analyzed promoter regions.

If `x` belongs to another supported class, the return format may vary.

Details

Methods should be implemented for specific object classes storing sequence data.

Examples

pfm1_path <- system.file("extdata", "pfm1.rds", package = "PscanR")
pfm1 <- readRDS(pfm1_path)
ps_seq_names(pfm1)
#>  [1] "NM_001078.4" "NM_000639.3" "NM_004433.5" "NM_001877.5" "NM_001561.6"
#>  [6] "NM_003039.3" "NM_000952.5" "NM_001993.5" "NM_014624.4" "NM_000567.3"
#> [11] "NM_001778.4" "NM_003005.4" "NM_000450.2" "NM_000963.4" "NM_000577.5"
#> [16] "NM_004591.3" "NM_015727.3" "NM_000575.5" "NM_000576.3" "NM_000579.4"
#> [21] "NM_002852.4" "NM_006225.4" "NM_005191.4" "NM_000584.4" "NM_003998.4"
#> [26] "NM_002994.5" "NM_000586.4" "NM_005429.5" "NM_002199.4" "NM_000758.4"
#> [31] "NM_002188.3" "NM_002198.3" "NM_000590.2" "NM_004355.4" "NM_002460.4"
#> [36] "NM_004233.4" "NM_000595.4" "NM_000594.4" "NM_002800.5" "NM_006290.4"
#> [41] "NM_000914.5" "NM_000367.5" "NM_003897.4" "NM_009588.1" "NM_172197.3"
#> [46] "NM_000593.6" "NM_000636.4" "NM_000600.5" "NM_002787.5" "NM_006528.4"