Save background statistics from a `PSMatrixList` object to a file

Saves background statistics (such as size, mean, and standard deviation) for each Position Weight Matrix in a `PSMatrixList` object to a specified file.

ps_write_bg_to_file(pfms, file)

Arguments

pfms

A `PSMatrixList` object of Position Weight Matrices representing transcription factor binding preferences, obtained, for example, from the JASPAR database. Each matrix should already include background statistics (e.g., computed from all promoter regions in the organism of study). Each element must be a `PSMatrix` object or coercible to one.

file

A character string specifying the path to the output file where the background statistics should be saved.

Value

None. It saves the given background statistics to the specified file in a tab-delimited format.

Details

This function first extracts background statistics from the input `pfms` using `ps_get_bg_table()`. It then writes the result to the specified file in a tab-delimited format. The output file begins with a header line: `[SHORT TFBS MATRIX]`. This header is used internally for compatibility with `ps_retrieve_bg_from_file()`, which expects it when reading.

The output does not include column names; the first column contains matrix identifiers, followed by `BG_SIZE`, `BG_MEAN`, and `BG_STDEV`.

This function uses example datasets located in the `extdata/` directory for demonstration purposes only. These files are not part of the core data used by the function. They can be accessed using `system.file()` as shown in the examples.

See also

Examples

# Load the example dataset for JASPAR2020 matrices collection 
# for vertebrates.
J2020_path <- system.file("extdata", "J2020.rds", package = "PscanR")
J2020 <- readRDS(J2020_path)
# File path to save the result
file_path <- "J2020_hg38_bg_stats.txt"

PSM1 <- PSMatrix(
  pfm = J2020[[1]],
  ps_bg_avg = 0.25,        
  ps_fg_avg = 0.5,         
  ps_bg_std_dev = 0.05,    
  ps_bg_size = 250L        
  )
PSM2 <- PSMatrix(
  pfm = J2020[[2]],
  ps_bg_avg = 0.25,        
  ps_fg_avg = 0.5,         
  ps_bg_std_dev = 0.05,    
  ps_bg_size = 250L        
  )
  
PSMatrixList_J2020 <- PSMatrixList(PSM1, PSM2)

# ps_write_bg_to_file(PSMatrixList_J2020, file_path)